auth done

+++ app/Http/Controllers/PostController.php
@@ -6,14 +6,14 @@ use App\Models\Like;
 use App\Models\Post;
 use App\Models\Tag;
 use Illuminate\Http\Request;
-use Illuminate\Session\Store;
+use Auth;
+use Gate;
 
 class PostController extends Controller
 {
     public function getIndex()
     {
         $posts = Post::orderBy('created_at', 'desc')->paginate(2);
-
         return view('blog.index', ['posts' => $posts]);
     }
 
@@ -25,7 +25,8 @@ class PostController extends Controller
 
     public function getPost($id)
     {
-        $post =  Post::find($id);
+//        $post =  Post::find($id);
+        $post = Post::where('id', $id)->with('likes')->first();
         return view('blog.post', ['post' => $post]);
     }
 
@@ -56,11 +57,15 @@ class PostController extends Controller
             'title' => 'required|min:5',
             'content' => 'required|min:10'
         ]);
+        $user = Auth::user();
+        if (!$user) {
+            return redirect()->back();
+        }
         $post = new Post([
             "title" => $request->input('title'),
             "content" => $request->input('content'),
         ]);
-        $post->save();
+        $user->posts()->save($post);
         $post->tags()->attach($request->input('tags') === null ? [] : $request->input('tags'));
         return redirect()->route('admin.index')->with('info', 'Post created, Title is: ' . $request->input('title'));
     }
@@ -73,6 +78,9 @@ class PostController extends Controller
         ]);
 
         $post = Post::find($request->input('id'));
+        if (Gate::denies('manipulate-post', $post)) {
+            return redirect()->back();
+        }
         $post->title = $request->input('title');
         $post->content = $request->input('content');
         $post->save();
@@ -84,7 +92,10 @@ class PostController extends Controller
 
     public function getAdminDelete($id)
     {
-        $post =  Post::find($id);
+        $post = Post::find($id);
+        if (Gate::denies('manipulate-post', $post)) {
+            return redirect()->back();
+        }
         $post->likes()->delete();
         $post->tags()->detach();
         $post->delete();
